requirements.txt
flask
langchain>=0.1.13
langchain-community>=0.0.32
transformers
sentence-transformers
torch
faiss-cpu
python-dotenv
gunicorn
gradio>=4.0.0